% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binned_binomial.R
\name{binned_binomial}
\alias{binned_binomial}
\title{Binned Binomial}
\usage{
binned_binomial(size_vect, prob_vect)
}
\arguments{
\item{size_vect}{A numeric vector for the sizes of the binomial distribution}

\item{prob_vect}{A numeric vector for the probabilities of the binomial distribution}
}
\value{
A numeric vector from the binomial distribution
}
\description{
Binned Binomial
}
\examples{
binned_binomial(c(10,20),c(0.5,0.6))
}
