% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binned_binomial.R
\name{binomial_size_prob}
\alias{binomial_size_prob}
\title{A Binomial Sample}
\usage{
binomial_size_prob(size_val, prob_val)
}
\arguments{
\item{size_val}{A numeric value for the size of the binomial distribution}

\item{prob_val}{A numeric value for the probability of the binomial distribution}
}
\value{
A numeric value from the binomial distribution
}
\description{
A function to generate a binomial sample
This is a simple wrapper around the rbinom function to be called
by the \code{binned_binomial} function
}
\examples{
binomial_size_prob(10,0.5)
}
