% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turelli_equilibria.R
\name{ci_stable}
\alias{ci_stable}
\title{Stable equilibrium in case of CI}
\usage{
ci_stable(F_val, mu_val, sh_val)
}
\arguments{
\item{F_val}{A numeric value greater than 0}

\item{mu_val}{A numeric value between 0 and 1}

\item{sh_val}{A numeric value between 0 and 1}
}
\value{
A numeric value between 0 and 1
}
\description{
Implements
\deqn{p^{\ast} = \frac{s_h + 1 - F + \sqrt{(s_h + 1 - F)^2 + 4s_h(F(1-\mu)-1)(1-F\mu)}}{2s_h(1 - F\mu)}}
}
\examples{
ci_stable(1.021,0.1,0.8)
}
