% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_freq_rf_rmu_ci.R
\name{infection_freq_rf_rmu_ci_iteration}
\alias{infection_freq_rf_rmu_ci_iteration}
\title{Iterate Random F and mu infection frequency model with CI}
\usage{
infection_freq_rf_rmu_ci_iteration(
  F_val_m,
  F_cv,
  mu_vect,
  bin_props,
  s_h,
  N_val,
  p_t_init = 0.4,
  n_iter = 10000,
  lwr_thresh = 1e-04
)
}
\arguments{
\item{F_val_m}{A numeric value for the mean fitness}

\item{F_cv}{A numeric value for the coefficient of variation for the fitness}

\item{mu_vect}{A numeric vector of transmission rates for each group}

\item{bin_props}{A numeric vector of proportions for each group}

\item{s_h}{A numeric value for the CI}

\item{N_val}{A numeric value for the total population size}

\item{p_t_init}{A numeric value between 0 and 1 for the initial infection frequency}

\item{n_iter}{A numeric value for the maximum number of iterations}

\item{lwr_thresh}{A numeric value for the lower threshold}
}
\value{
A numeric vector of infection frequencies
}
\description{
Iterate Random F and mu infection frequency model with CI
}
\examples{
ptlt <- infection_freq_rf_rmu_ci_iteration(1.2,0.1,c(0.01,0.9),c(0.99,0.01),0.1,1000)
plot_a_simulation(ptlt,alpha=0.8)
extract_stats(ptlt)
}
