% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_freq_rf_rmu_ci.R
\name{infection_freq_rf_rmu_ci_sim}
\alias{infection_freq_rf_rmu_ci_sim}
\title{Main simulation for random F and mu model with CI}
\usage{
infection_freq_rf_rmu_ci_sim(
  rep_num,
  F_val_m,
  F_cv,
  mu_vect,
  bin_props,
  s_h,
  N_val,
  p_t_init = 0.4,
  ...
)
}
\arguments{
\item{rep_num}{A numeric value for the number of replications}

\item{F_val_m}{A numeric value for the mean fitness}

\item{F_cv}{A numeric value for the coefficient of variation for the fitness}

\item{mu_vect}{A numeric vector of transmission rates for each group}

\item{bin_props}{A numeric vector of proportions for each group}

\item{s_h}{A numeric value for the CI}

\item{N_val}{A numeric value for the total population size}

\item{p_t_init}{A numeric value between 0 and 1 for the initial infection frequency}

\item{...}{Additional arguments}
}
\value{
A tibble of simulation results
}
\description{
This function simulates the infection frequency
for random fitness and mu values
}
\examples{
infection_freq_rf_rmu_ci_sim(1,1.2,0.1,c(0.01,0.9),c(0.99,0.01),0.1,1000)
}
