% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_freq_rmu.R
\name{infection_freq_rmu}
\alias{infection_freq_rmu}
\title{Random mu infection frequency model}
\usage{
infection_freq_rmu(p_t, F_val, mu_vect, bin_props, N_val)
}
\arguments{
\item{p_t}{A numeric value between 0 and 1}

\item{F_val}{A numeric value for the fitness}

\item{mu_vect}{A numeric vector of transmission rates for each group}

\item{bin_props}{A numeric vector of proportions for each group}

\item{N_val}{A numeric value for the total population size}
}
\value{
A numeric value for the infection frequency
}
\description{
Random mu infection frequency model
}
\examples{
infection_freq_rmu(0.5,1,c(0.1,0.2),c(0.5,0.5),100)
}
