% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_freq_rmu_ci.R
\name{infection_freq_rmu_ci}
\alias{infection_freq_rmu_ci}
\title{Random mu infection frequency model with CI}
\usage{
infection_freq_rmu_ci(p_t, F_val, mu_vect, bin_props, s_h, N_val)
}
\arguments{
\item{p_t}{A numeric value between 0 and 1}

\item{F_val}{A numeric value for the fitness}

\item{mu_vect}{A numeric vector of transmission rates for each group}

\item{bin_props}{A numeric vector of proportions for each group}

\item{s_h}{A numeric value for the CI}

\item{N_val}{A numeric value for the total population size}
}
\value{
A numeric value for the infection frequency
}
\description{
Random mu infection frequency model with CI
}
\examples{
infection_freq_rmu_ci(0.4,1.2,c(0.01,0.8),c(0.99,0.01),0.1,1000)
}
