% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_freq_rmu_ci.R
\name{infection_freq_rmu_ci_sim}
\alias{infection_freq_rmu_ci_sim}
\title{Main simulation for fixed F but random mu model with CI}
\usage{
infection_freq_rmu_ci_sim(
  rep_num,
  F_val,
  mu_vect,
  bin_props,
  s_h,
  N_val,
  p_t_init = 0.4,
  ...
)
}
\arguments{
\item{rep_num}{A numeric value for the number of replications}

\item{F_val}{A numeric value for the fitness}

\item{mu_vect}{A numeric vector of transmission rates for each group}

\item{bin_props}{A numeric vector of proportions for each group}

\item{s_h}{A numeric value for the CI}

\item{N_val}{A numeric value for the total population size}

\item{p_t_init}{A numeric value between 0 and 1}

\item{...}{Additional arguments}
}
\value{
A tibble of the simulation results
}
\description{
This function simulates the infection frequency
for a fixed fitness value but random mu values
but includes CI
}
\examples{
infection_freq_rmu_ci_sim(100,1.2,c(0.01,0.8),c(0.99,0.01),0.1,1000)
}
