% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_freq_rmu.R
\name{infection_freq_rmu_sim}
\alias{infection_freq_rmu_sim}
\title{Main simulation for fixed F but random mu model}
\usage{
infection_freq_rmu_sim(
  rep_num,
  F_val,
  mu_vect,
  bin_props,
  N_val,
  p_t_init = 0.4,
  ...
)
}
\arguments{
\item{rep_num}{A numeric value for the number of replications}

\item{F_val}{A numeric value for the fitness}

\item{mu_vect}{A numeric vector of transmission rates for each group}

\item{bin_props}{A numeric vector of proportions for each group}

\item{N_val}{A numeric value for the total population size}

\item{p_t_init}{A numeric value between 0 and 1 for the initial infection frequency}

\item{...}{Additional arguments}
}
\value{
A tibble with the simulation results
}
\description{
This function simulates the infection frequency
for a fixed fitness value but random mu values
}
\examples{
infection_freq_rmu_sim(1,1.05,c(0.01,0.9),c(0.99,0.01),1000)
}
