% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turelli_equilibria.R
\name{mu_F_equality}
\alias{mu_F_equality}
\title{Curve where no-CI equilibria can be maintained}
\usage{
mu_F_equality(F_val)
}
\arguments{
\item{F_val}{A numeric value greater than 0}
}
\value{
A numeric value
}
\description{
Implements
\deqn{\mu = 1 - \frac{1}{F}}
}
\examples{
mu_F_equality(1.1)
}
