% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turelli_equilibria.R
\name{no_ci}
\alias{no_ci}
\title{Stable equilibrium in case of no CI}
\usage{
no_ci(F_val, mu_val)
}
\arguments{
\item{F_val}{A numeric value greater than 0}

\item{mu_val}{A numeric value between 0 and 1}
}
\value{
A numeric value between 0 and 1
}
\description{
Implements
\deqn{p^{\ast} = 1 - \frac{\mu F}{F - 1}}
}
\examples{
no_ci(1.2,0.1)
}
