% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_a_simulation.R
\name{plot_a_simulation}
\alias{plot_a_simulation}
\title{Plot a simulation}
\usage{
plot_a_simulation(p_t_vec, threshold = 1e-04, alpha_val = 0.2)
}
\arguments{
\item{p_t_vec}{A numeric vector}

\item{threshold}{A numeric value}

\item{alpha_val}{A numeric value}
}
\value{
A ggplot object
}
\description{
Plots a simulation
}
\examples{
plot_a_simulation(runif(1000))
}
