% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_mu_ci_sim_data.R
\docType{data}
\name{rand_mu_ci_045_sim_data_1}
\alias{rand_mu_ci_045_sim_data_1}
\title{Simulation data for random mu model with CI.}
\format{
\subsection{\code{rand_mu_ci_045_sim_data_1}}{

A data frame with 27,500 rows and 13 columns:
\describe{
\item{rep_num}{Repitition number}
\item{F_val}{F value}
\item{bin_props}{group proportion for mu}
\item{s_h}{CI value}
\item{N_val}{Population size}
\item{p_t_median}{median infection proportion}
\item{p_t_mean}{mean infection proportion}
\item{p_t_var}{variance of infection proportion}
\item{p_t_min}{minimum infection proportion}
\item{p_t_max}{maximum infection proportion}
\item{gen_steps}{number of generations}
\item{final_p_t}{final infection proportion}
...
}
}
}
\source{
Simulated from a model.
}
\usage{
rand_mu_ci_045_sim_data_1
}
\description{
Model-simulated data for multi-binomial distributed mu
Single mu group scenario with CI at 0.45.
}
\keyword{datasets}
