% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turelli_2022_fig_02_data.R
\docType{data}
\name{turelli_2022_fig_02_data_1}
\alias{turelli_2022_fig_02_data_1}
\title{Turelli et al. 2022 Figure 2 F_m = 1.021 data}
\format{
\subsection{\code{turelli_2022_fig_02_data_1}}{

A data frame with 200 rows and 4 columns:
\describe{
\item{F_val_m}{Median F value}
\item{sh_val}{CI value}
\item{repeat_num}{Repitition number}
\item{t_f}{Persistence time}
...
}
}
}
\source{
Simulated from a model following Turelli et al. 2022
}
\usage{
turelli_2022_fig_02_data_1
}
\description{
Model-simulated data for Figure 2 of Turelli et al. 2022.
This data corresponds to the case where the mean value of
F is 1.021 for the log-normal distribution of F.
}
\keyword{datasets}
