% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turelli_model.R
\name{turelli_iteration}
\alias{turelli_iteration}
\title{Turelli Iteration}
\usage{
turelli_iteration(p_val, F_val, mu_val, sh_val, N)
}
\arguments{
\item{p_val}{A numeric value between 0 and 1}

\item{F_val}{A numeric value greater than 0}

\item{mu_val}{A numeric value between 0 and 1}

\item{sh_val}{A numeric value between 0 and 1}

\item{N}{A numeric value greater than 0}
}
\value{
A numeric value between 0 and 1
}
\description{
Implements the basic iteration for the Turelli model
assuming a finite population. This calls the function \code{turelli_model}.
}
\examples{
turelli_iteration(0.4,1.021,0.1,0.0,1000)
}
