% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turelli_model.R
\name{turelli_model}
\alias{turelli_model}
\title{Turelli Model Function}
\usage{
turelli_model(p_t, F_val, mu_val, sh_val = 0)
}
\arguments{
\item{p_t}{A numeric value between 0 and 1}

\item{F_val}{A numeric value greater than 0}

\item{mu_val}{A numeric value between 0 and 1}

\item{sh_val}{A numeric value between 0 and 1}
}
\value{
A numeric value between 0 and 1
}
\description{
Turelli Model Function
}
\examples{
turelli_model(0.4,1.021,0.1)
}
