% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turelli_model.R
\name{turelli_simulation}
\alias{turelli_simulation}
\title{Simulation of Turelli model}
\usage{
turelli_simulation(
  p_0,
  F_val,
  mu_val,
  sh_val,
  N,
  max_iter = 10^4,
  thresh = 10^(-5)
)
}
\arguments{
\item{p_0}{A numeric value between 0 and 1}

\item{F_val}{A numeric value greater than 0}

\item{mu_val}{A numeric value between 0 and 1}

\item{sh_val}{A numeric value between 0 and 1}

\item{N}{A numeric value greater than 0}

\item{max_iter}{A numeric value greater than 0}

\item{thresh}{A numeric value between 0 and 1}
}
\value{
A numeric vector with the frequency of infected over time
}
\description{
Simulates the Turelli model for a given number of generations
or until the frequency is below a given threshold.
}
\examples{
turelli_simulation(0.4,1.021,0.1,0.0,1000)
}
