% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turelli_model_rf.R
\name{turelli_simulation_rf_tf}
\alias{turelli_simulation_rf_tf}
\title{Simulation of Turelli et al. 2022 model}
\usage{
turelli_simulation_rf_tf(
  p_0,
  F_val_m,
  F_cv,
  mu_val,
  sh_val,
  N,
  max_iter = 10^8,
  thresh = 10^(-5)
)
}
\arguments{
\item{p_0}{A numeric value between 0 and 1}

\item{F_val_m}{A numeric value greater than 0}

\item{F_cv}{A numeric value greater than 0}

\item{mu_val}{A numeric value between 0 and 1}

\item{sh_val}{A numeric value between 0 and 1}

\item{N}{A numeric value greater than 0}

\item{max_iter}{A numeric value greater than 0}

\item{thresh}{A numeric value between 0 and 1}
}
\value{
A numeric vector with the frequency of infected over time
}
\description{
Simulates the Turelli et al. 2022 model for a given number of generations
or until the frequency is below a given threshold. This model
incorporates a randomly fluctuating fitness value.
}
\examples{
turelli_simulation_rf_tf(0.4,1.021,0.4,0.1,0.0,1000)
}
