% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turelli_model.R
\name{turelli_model}
\alias{turelli_model}
\title{Turelli Model Function}
\usage{
turelli_model(p_t, F_val, mu_val, sh_val = 0)
}
\arguments{
\item{p_t}{A numeric value between 0 and 1}

\item{F_val}{A numeric value greater than 0}

\item{mu_val}{A numeric value between 0 and 1}

\item{sh_val}{A numeric value between 0 and 1}
}
\value{
A numeric value between 0 and 1
}
\description{
Implements the iteration

\deqn{p_{t+1} = \frac{p_t F (1 - \mu)}{1 + p_t (F - 1 - s_h) + p_t^2 s_h (1 - \mu F)}}

This is the basic deterministic model.
}
\examples{
turelli_model(0.4,1.021,0.1)
}
